remove gdk_rpc
