+++ subprojects/gdk_rust/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "gdk-rust"
 version = "0.1.0"
-authors = ["Nadav Ivgi <nadav@shesek.info>"]
+authors = ["Riccardo Casatta <riccardo@casatta.it>", "Nadav Ivgi <nadav@shesek.info>"]
 license = "(MIT)"
 edition = "2018"
 
@@ -10,30 +10,22 @@ crate-type = ["staticlib"]
 
 [features]
 android_log = ["android_logger"]
-stderr_logger = ["stderrlog"]
 
 [dependencies]
 bitcoin = { version = "0.23", features = [ "use-serde" ] }
 rand = "0.7"
 gdk-electrum = { path = "gdk_electrum", features = ["android_log"] }
-# gdk-rpc = { path = "gdk_rpc" }
 gdk-common = { path = "gdk_common" }
 serde = "1.0"
 serde_json = "1.0"
 serde_derive = "1.0"
 libc = "0.2"
-lazy_static = "1.3"
-failure = "0.1"
 hex = "0.3"
 dirs = "1.0"
 chrono = "0.4"
 log = "0.4"
-hyper = "0.10"
-hyper-socks = "0.4"
-stderrlog = { version = "0.4", optional = true }
 android_logger = { version = "0.8.6", optional = true }
 url = "1.7.2"
-backtrace = "0.3"
 ureq = { version = "1.0.0", features = ["json"] }
 
 [dev-dependencies]
