--- subprojects/gdk_rust/README.md
@@ -1,4 +1,0 @@
-# GDK-RPC
-
-GDK-RPC library allows users to use the Green mobile/cli/desktop apps
-with a Bitcoin or Liquid full node including pruned.
