--- subprojects/gdk_rust/gdk_common/src/constants.rs
@@ -1,15 +1,0 @@
-pub const GA_OK: i32 = 0;
-pub const GA_ERROR: i32 = -1;
-pub const GA_RECONNECT: i32 = -2;
-pub const GA_SESSION_LOST: i32 = -3;
-pub const GA_TIMEOUT: i32 = -4;
-pub const GA_NOT_AUTHORIZED: i32 = -5;
-
-pub const GA_TRUE: u32 = 1;
-pub const GA_FALSE: u32 = 0;
-
-pub const GA_MEMO_USER: u32 = 0;
-
-pub const SAT_PER_BTC: f64 = 100_000_000.0;
-pub const SAT_PER_MBTC: f64 = 100_000.0;
-pub const SAT_PER_BIT: f64 = 100.0;
