+++ subprojects/gdk_rust/gdk_common/src/lib.rs
@@ -1,15 +1,10 @@
-#[macro_use]
-extern crate serde_json;
-
 pub mod be;
-pub mod constants;
 pub mod error;
 pub mod mnemonic;
 pub mod model;
 pub mod network;
 pub mod password;
 pub mod session;
-pub mod util;
 pub mod wally;
 pub mod scripts;
