--- subprojects/gdk_rust/gdk_rpc/Cargo.toml
@@ -1,52 +1,0 @@
-[package]
-name = "gdk-rpc"
-version = "0.1.0"
-authors = ["Nadav Ivgi <nadav@shesek.info>"]
-license = "(MIT)"
-edition = "2018"
-
-[features]
-liquid = ["elements"]
-android_log = ["android_logger"]
-stderr_logger = ["stderrlog"]
-
-[dependencies]
-bitcoin = { version = "0.23", features = [ "use-serde" ] }
-bitcoin_hashes = { version = "0.7", features = [ "serde" ] }
-rand = "0.7"
-serde = "1.0"
-serde_json = "1.0"
-serde_derive = "1.0"
-lazy_static = "1.3"
-failure = "0.1"
-hex = "0.3"
-dirs = "1.0"
-chrono = "0.4"
-log = "0.4"
-hyper = "0.10"
-hyper-socks = "0.4"
-stderrlog = { version = "0.4", optional = true }
-android_logger = { version = "0.8.6", optional = true }
-url = "1.7.2"
-backtrace = "0.3"
-gdk-common = { path = "../gdk_common" }
-
-
-[dependencies.jsonrpc]
-git = "https://github.com/jb55/rust-jsonrpc"
-rev = "5e4dd4817641aa2f5b9981b4ea7f3bca52e0b3d9"
-
-[dependencies.bitcoincore-rpc]
-git = "https://github.com/jb55/rust-bitcoincore-rpc"
-rev = "06c9e44d526b285c2218075ebb4bf539957f21f1"
-
-[dependencies.elements]
-version = "0.7.1"
-optional = true
-
-[profile.release]
-lto = true
-panic = 'abort'
-opt-level = 'z'
-codegen-units = 1
-incremental = false
