--- subprojects/gdk_rust/gdk_rpc/src/lib.rs
@@ -1,17 +1,0 @@
-#[macro_use]
-extern crate serde_json;
-
-#[macro_use]
-extern crate log;
-
-#[macro_use]
-extern crate serde;
-
-#[macro_use]
-pub mod errors;
-
-pub mod coins;
-pub mod network;
-pub mod session;
-pub mod settings;
-pub mod wallet;
